add ui pagefb
