+++ LOMSUI/Services/ApiService.cs
@@ -84,7 +84,7 @@ namespace LOMSUI.Services
         {
             try
             {
-                var response = await _httpClient.PutAsync($"{BASE_URL}/update-token-facebook?pageid={pageid}", null);
+                var response = await _httpClient.PutAsync($"{BASE_URL}/update-page-id?pageId={pageid}", null);
                 var responseBody = await response.Content.ReadAsStringAsync();
                 Console.WriteLine($"Update Page Response: {responseBody}");
